requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
psycopg2-binary
python-dotenv==1.0.1
openai==1.57.2
requests==2.32.3
python-multipart==0.0.20
sqlmodel==0.0.22
reportlab==4.2.5
passlib[bcrypt]>=1.7.4
bcrypt==4.1.2
PyJWT>=2.8.0
beautifulsoup4>=4.12.0
feedparser>=6.0.10
